requirements.txt
# =============================================================================
# RIDE CLI - Requirements
# =============================================================================
# This file mirrors pyproject.toml for reproducible development environments.
# Install all dependencies:   pip install -r requirements.txt
# Install the package itself: pip install -e ".[dev]"
# =============================================================================

# --- Core engine -------------------------------------------------------------
polars-lts-cpu>=0.20.0
numpy>=1.26.0
pyarrow>=14.0.0

# --- CLI & terminal UI -------------------------------------------------------
rich>=13.0.0
click>=8.0.0
pyfiglet==0.8.post1
termcolor>=2.0.0
blessed==1.19.1
plotext==5.2.8

# --- Data formats ------------------------------------------------------------
fastparquet>=2024.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
nbformat>=5.9.0
pandas>=2.0.0          # Fallback for Excel loading (.xlsx/.xls) only

# --- ML / statistics ---------------------------------------------------------
scikit-learn>=1.3.0
scipy>=1.11.0
joblib>=1.2.0

# --- Configuration -----------------------------------------------------------
pydantic>=2.0.0

# --- Progress ----------------------------------------------------------------
tqdm>=4.64.0

# --- Optional: SQL (advanced extras) ----------------------------------------
# duckdb>=0.9.0        # Uncomment to enable SQL query mode (--sql flag)
#                      # Or install with: pip install "ride-cli[advanced]"

# --- Optional: Visualization (viz extras) ------------------------------------
# matplotlib>=3.7.0,<4.0.0
# seaborn>=0.12.0,<1.0.0
#                      # Or install with: pip install "ride-cli[viz]"

# --- Development only --------------------------------------------------------
pytest>=7.3.0
pytest-cov>=4.0.0
ruff
black
isort
mypy
pre-commit
twine
wheel
